CLI: Refine `site stop --all` implementation (#2374)
* Moved logic inside stop command and added flag

* Simplify `site stop --all` implementation

* Use `site stop --all` from Studio

* Fix tests

* Address review comments

---------
